src/sync/requirements.txt
requests>=2.33.0
psycopg2-binary>=2.9.10
